https://www.reddit.com/r/reinforcementlearning/comments/8bj70c/d_what_exactly_is_the_difference_between_onpolicy/dx8lssf/?st=jfwi55ne&sh=d647c73e
Just look at what essays I've been writting here. This has been essentially my work for the day.

Ah, fuck it. I cannot muster the will. When I get right down to it, I just believe in PG and DQN so little.

Let me take a detour. It is like my inner sense is telling me to just predict the dynamics and then plan off that. That is where RL is moving towards.

Now I'll admit, I've skipped the model based RL lectures entirely. How about I watch them now.

Eventually I will gather the will to do this, little by little.

I just cannot help it.

When I make a mistake in programming there is like a sixth sense telling me that I've gone wrong somewhere. And now it is telling me that I need more. More, more...if I thought this would be enough, I'd have done this long ago.

I really want to solve the whole issue of making money so I do not want to be in the exploration stage forever, but on the other hand I do not want to neglect fundamentals.

So let me educate myself a bit more.

It is not like there is much to left to go apart from those model based learning lectures. Tomorrow I will try and get complete focus.

But this is classical me. I do not feel like moving forward unless I have a complete plan. One of the reasons why I decided to fuck off for 3 years and make a language and a ML library is so the RL field has time to grow into what I want it to be.

So far the progress has been slower than I anticipate.

The fact that I have to explain why scalar rewards are a bad idea is not annoying, but scary to me. It makes me fear for my future.

Forget this for a while. Let me go through the videos. I have absolutely no idea when I will get a capable player going, but I'll try my best either way.

I'll just worry and anguish until I get it done.

+++ Games/Tests.fs
@@ -53,7 +53,7 @@ let poker4 =
     """
 inl log _ _ = ()
 open Poker log
-inl a = {reply=reply_q {init=5f64; learning_rate=0.03f64; num_players=2}; name="One"; trace=term_cast (inl _ -> ()) float64}
+inl a = {reply=reply_q {init=5f64; learning_rate=0.03f64; num_players=2}; name="One"; trace=term_cast (inl _ -> ()) int64}
 inl b = {reply=reply_random; name="Two"}
 Loops.for {from=0; near_to=100; body=inl {i} ->
     Timer.time_it (string_format "iteration {0}" i)
